# api/openapi.runtime.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "JobPilot Platform API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/admin/schedules": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "List Schedules",
    "operationId": "list_schedules_api_v1_admin_schedules_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ScheduleOut"
         },
         "type": "array",
         "title": "Response List Schedules Api V1 Admin Schedules Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/admin/schedules/{task_key}": {
   "patch": {
    "tags": [
     "admin"
    ],
    "summary": "Update Schedule",
    "operationId": "update_schedule_api_v1_admin_schedules__task_key__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "task_key",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Task Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SchedulePatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ScheduleOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/schedules/{task_key}/run": {
   "post": {
    "tags": [
     "admin"
    ],
    "summary": "Run Schedule Now",
    "operationId": "run_schedule_now_api_v1_admin_schedules__task_key__run_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "task_key",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Task Key"
      }
     }
    ],
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/runs": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "List Runs",
    "operationId": "list_runs_api_v1_admin_runs_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "task_key",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Task Key"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "default": 50,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/RunOut"
         },
         "title": "Response List Runs Api V1 Admin Runs Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/connectors": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "List Connectors",
    "operationId": "list_connectors_api_v1_admin_connectors_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ConnectorOut"
         },
         "type": "array",
         "title": "Response List Connectors Api V1 Admin Connectors Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/admin/connectors/{connector_id}": {
   "patch": {
    "tags": [
     "admin"
    ],
    "summary": "Update Connector",
    "operationId": "update_connector_api_v1_admin_connectors__connector_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "connector_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Connector Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ConnectorPatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ConnectorOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
...
   "ConnectorOut": {
    "properties": {
     "connector_id": {
      "type": "string",
      "title": "Connector Id"
     },
     "display_name": {
      "type": "string",
      "title": "Display Name"
     },
     "compliance_mode": {
      "type": "string",
      "title": "Compliance Mode"
     },
     "enabled": {
      "type": "boolean",
      "title": "Enabled"
     },
     "rate_limit_per_min": {
      "type": "integer",
      "title": "Rate Limit Per Min"
     }
    },
    "type": "object",
    "required": [
     "connector_id",
     "display_name",
     "compliance_mode",
     "enabled",
     "rate_limit_per_min"
    ],
    "title": "ConnectorOut"
   },
   "ConnectorPatch": {
    "properties": {
     "enabled": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Enabled"
     },
     "rate_limit_per_min": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 600.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Rate Limit Per Min"
     }
    },
    "type": "object",
    "title": "ConnectorPatch"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "RunOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "task_key": {
      "type": "string",
      "title": "Task Key"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "started_at": {
      "title": "Started At"
     },
     "finished_at": {
      "anyOf": [
       {},
       {
        "type": "null"
       }
      ],
      "title": "Finished At"
     },
     "stats": {
      "additionalProperties": true,
      "type": "object",
      "title": "Stats",
      "default": {}
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     }
    },
    "type": "object",
    "required": [
     "id",
     "task_key",
     "status",
     "started_at"
    ],
    "title": "RunOut"
   },
   "ScheduleOut": {
    "properties": {
     "key": {
      "type": "string",
      "title": "Key"
     },
     "cron": {
      "type": "string",
      "title": "Cron"
     },
     "timezone": {
      "type": "string",
      "title": "Timezone"
     },
     "enabled": {
      "type": "boolean",
      "title": "Enabled"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     }
    },
    "type": "object",
    "required": [
     "key",
     "cron",
     "timezone",
     "enabled"
    ],
    "title": "ScheduleOut"
   },
   "SchedulePatch": {
    "properties": {
     "cron": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 64,
        "minLength": 9
       },
       {
        "type": "null"
       }
      ],
      "title": "Cron"
     },
     "enabled": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Enabled"
     }
    },
    "type": "object",
    "title": "SchedulePatch"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}